Actualizacion de controllers
