+++ loscobresdesarrollophp/controllers/AuthenticationController.php
@@ -1,0 +1,5 @@
+<?php
+class AuthenticationController {
+    // Placeholder for authentication-related methods
+}
+?>
