+++ loscobresdesarrollophp/controllers/ViewController.php
@@ -1,0 +1,5 @@
+<?php
+class ViewController {
+    // Placeholder for view-related methods
+}
+?>
